Move localRPCTransport from transport.go to transport_test.go
